backend/requirements.txt
# Cash Flow Analysis System - Backend Dependencies
# Install with: pip install -r requirements.txt

# ===== WEB FRAMEWORK =====
Flask>=2.0.0
flask-cors>=3.0.10

# ===== ENVIRONMENT VARIABLES =====
python-dotenv>=0.19.0

# ===== DATA PROCESSING =====
pandas>=1.3.0
numpy>=1.21.0
openpyxl>=3.0.0

# ===== DATABASE =====
mysql-connector-python>=8.2.0

# ===== AI AND OPENAI =====
openai>=1.12.0
sentence-transformers>=2.2.2

# ===== MACHINE LEARNING =====
scikit-learn>=1.4.0
xgboost>=2.0.3
scipy>=1.7.0

# ===== PDF GENERATION =====
reportlab>=3.6.0

# ===== HTTP REQUESTS =====
requests>=2.25.0

# ===== DATE/TIME UTILITIES =====
python-dateutil>=2.8.0
pytz>=2024.1

# ===== TYPE HINTS (Python < 3.9) =====
typing-extensions>=4.9.0

# ===== UTILITIES =====
psutil>=5.8.0

# ===== OPTIONAL: DEEP LEARNING (if needed) =====
# tensorflow>=2.15.0
# tf-keras>=2.20.0

# ===== OPTIONAL: PRODUCTION SERVER =====
# gunicorn>=20.1.0
# waitress>=2.0.0

# ===== OPTIONAL: DEVELOPMENT =====
# pytest>=6.0.0
# black>=21.0.0
# flake8>=3.9.0
